requirements.txt
streamlit
openai
python-dotenv
psycopg2-binary
google-generativeai
numpy
pandas
plotly
catsim